+++ cercle-5/miniRT/srcs/gui/editor/panel/transform/mesh.c
@@ -6,7 +6,7 @@
 /*   By: abdoali <abdoali@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/07 00:00:00 by abdoali           #+#    #+#             */
-/*   Updated: 2026/04/01 12:56:10 by abdoali          ###   ########.fr       */
+/*   Updated: 2026/04/03 13:41:46 by abdoali          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 static void	mesh_vertex_sync(t_mesh *m, t_mesh_sync s)
 {
-	int		vi;
+	size_t	vi;
 	t_vec3	local;
 
 	vi = 0;
-	while ((size_t)vi < m->vertex_count)
+	while (vi < m->vertex_count)
 	{
 		local = vec3_sub(m->edit_snap_verts[vi].pos, s.piv);
 		local = mat4_mul_pos(s.sr, local);
@@ -30,7 +30,7 @@ static void	mesh_vertex_sync(t_mesh *m, t_mesh_sync s)
 	}
 	m->bbox = aabb_create_empty();
 	vi = 0;
-	while ((size_t)vi < m->vertex_count)
+	while (vi < m->vertex_count)
 		aabb_expand_point(&m->bbox, m->vertices[vi++].pos);
 	mesh_build_bvh(m);
 }
@@ -40,7 +40,7 @@ void	mesh_transform_sync(t_gui *gui)
 	t_mesh_group	*g;
 	t_scene			*sc;
 	t_mat4			sr;
-	int				si;
+	size_t			si;
 
 	if (!gui->selection.active || gui->selection.type != TYPE_MESH)
 		return ;
@@ -54,7 +54,7 @@ void	mesh_transform_sync(t_gui *gui)
 	sr = mat4_mul(mat4_rotation(g->transform.rotation),
 			mat4_scaling(g->transform.scale));
 	si = 0;
-	while ((size_t)si < g->sub_count)
+	while (si < g->sub_count)
 	{
 		if (sc->meshes[g->start + si].edit_snap_verts)
 			mesh_vertex_sync(&sc->meshes[g->start + si], (t_mesh_sync){sr,
