+++ cercle-5/miniRT/srcs/primitives/objects/mesh/adder.c
@@ -6,7 +6,7 @@
 /*   By: abdoali <abdoali@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/30 21:35:00 by abdoali           #+#    #+#             */
-/*   Updated: 2026/04/03 11:39:11 by abdoali          ###   ########.fr       */
+/*   Updated: 2026/04/03 13:51:41 by abdoali          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,17 +43,17 @@ bool	scene_add_group(t_scene *scene, t_mesh_group g)
 
 static bool	process_resource_items(t_scene *scene, t_mesh_resource *res)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
-	while ((size_t)i < res->mesh_count)
+	while (i < res->mesh_count)
 	{
 		if (!scene_add_mesh(scene, res->meshes[i]))
 			return (false);
 		i++;
 	}
 	i = 0;
-	while ((size_t)i < res->group_count)
+	while (i < res->group_count)
 	{
 		if (!scene_add_group(scene, res->groups[i]))
 			return (false);
@@ -65,7 +65,7 @@ static bool	process_resource_items(t_scene *scene, t_mesh_resource *res)
 bool	scene_add_objects(t_scene *scene, const char *path)
 {
 	t_mesh_resource	res;
-	int				start;
+	size_t			start;
 
 	if (!scene || !path)
 		return (false);
@@ -74,7 +74,7 @@ bool	scene_add_objects(t_scene *scene, const char *path)
 	start = scene->mesh_count;
 	if (!process_resource_items(scene, &res))
 		return (false);
-	if (scene->mesh_count > (size_t)start)
+	if (scene->mesh_count > start)
 		scene_add_group_for_subs(scene, path, start);
 	return (true);
 }
