+++ cercle-5/miniRT/srcs/primitives/objects/mesh/intersect/traverse/leaf.c
@@ -6,7 +6,7 @@
 /*   By: abdoali <abdoali@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/15 00:00:00 by abdoali           #+#    #+#             */
-/*   Updated: 2026/04/01 19:15:00 by abdoali          ###   ########.fr       */
+/*   Updated: 2026/04/03 13:46:37 by abdoali          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,10 +17,10 @@ void	process_mesh_leaf(t_mesh *mesh, t_mbvh_node *node, const t_ray *ray,
 		t_trace *trace)
 {
 	t_leaf	leaf;
-	int		i;
+	size_t	i;
 
 	i = 0;
-	while ((size_t)i < node->count)
+	while (i < node->count)
 	{
 		PROF_INC(g_mesh_tri_tests);
 		leaf.tri = node->left_or_first + i;
