+++ cercle-5/miniRT/srcs/primitives/objects/mesh/trace/occlude/leaf.c
@@ -6,7 +6,7 @@
 /*   By: abdoali <abdoali@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/15 00:00:00 by abdoali           #+#    #+#             */
-/*   Updated: 2026/04/01 19:12:00 by abdoali          ###   ########.fr       */
+/*   Updated: 2026/04/03 13:46:37 by abdoali          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,12 +15,12 @@
 bool	leaf_occluded(t_mesh *mesh, t_mbvh_node *node, const t_ray *ray,
 		double dist)
 {
-	int		i;
-	int		tri;
+	size_t	i;
+	size_t	tri;
 	double	t;
 
 	i = 0;
-	while ((size_t)i < node->count)
+	while (i < node->count)
 	{
 		tri = node->left_or_first + i;
 		if (intersect_tri_precomp(ray, &mesh->tri_cache[tri], &t, NULL)
